+++ tools/rvlq/CLI/registry.h
@@ -34,6 +34,7 @@ namespace registry {
 			Input,
 			Output,
 			TagDelimiter,
+			UseCurrentDir,
 			CommentMarker,
 			TextMarker,
 			FileMarker,